refactor: 会話エントリの表示責務を分離する
ConversationEntry に集まっていたヘッダー表示、本文描画、編集 UI の責務を分割し、会話 UI の見通しを改善する。
