+++ src/app/(authed)/_components/ConversationEntry.tsx
@@ -1,32 +1,16 @@
 'use client';
 
-import { MoreVert } from '@mui/icons-material';
-import DeleteIcon from '@mui/icons-material/Delete';
-import {
-  Alert,
-  Avatar,
-  Box,
-  Chip,
-  IconButton,
-  Menu,
-  MenuItem,
-  Paper,
-  Snackbar,
-  Stack,
-  TextField,
-  Typography,
-} from '@mui/material';
-import { alpha } from '@mui/material/styles';
+import { Alert, Box, Snackbar } from '@mui/material';
 import type { MouseEvent } from 'react';
 import { useState } from 'react';
-import Markdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import { formatString } from '@/generic/DateFormatter';
 import type {
   ConversationActionResult,
   ConversationCapabilities,
   ConversationEntryViewModel,
 } from './conversationTypes';
+import ConversationEntryBody from './ConversationEntryBody';
+import ConversationEntryEditor from './ConversationEntryEditor';
+import ConversationEntryHeader from './ConversationEntryHeader';
 
 type Props = {
   entry: ConversationEntryViewModel;
@@ -56,13 +40,20 @@ const ConversationEntry = ({
     message: string;
   }>({ open: false, message: '' });
 
-  const isAdmin = entry.authorRole === 'ADMINISTRATOR';
-  const showMenuTrigger =
-    capabilities.actionTrigger === 'menu' && (entry.canEdit || entry.canDelete);
-  const showDeleteTrigger =
-    capabilities.actionTrigger === 'icon' && entry.canDelete;
-
   const showError = (message: string) => setSnackbar({ open: true, message });
+  const closeSnackbar = () => setSnackbar((prev) => ({ ...prev, open: false }));
+  const handleOpenMenu = (event: MouseEvent<HTMLElement>) =>
+    setAnchorEl(event.currentTarget);
+  const handleCloseMenu = () => setAnchorEl(undefined);
+  const handleStartEditing = () => {
+    setDraftBody(entry.body);
+    setIsEditing(true);
+    handleCloseMenu();
+  };
+  const handleCancelEditing = () => {
+    setDraftBody(entry.body);
+    setIsEditing(false);
+  };
 
   const handleDelete = async () => {
     if (!onDelete) {
@@ -75,7 +66,7 @@ const ConversationEntry = ({
     } else if (!result.success) {
       showError('不明なエラーが発生しました。もう一度お試しください。');
     }
-    setAnchorEl(undefined);
+    handleCloseMenu();
   };
 
   const handleUpdate = async () => {
@@ -102,158 +93,34 @@ const ConversationEntry = ({
       <Snackbar
         open={snackbar.open}
         autoHideDuration={6000}
-        onClose={() => setSnackbar((prev) => ({ ...prev, open: false }))}
+        onClose={closeSnackbar}
         anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
       >
-        <Alert
-          severity="error"
-          onClose={() => setSnackbar((prev) => ({ ...prev, open: false }))}
-        >
+        <Alert severity="error" onClose={closeSnackbar}>
           {snackbar.message}
         </Alert>
       </Snackbar>
       <Box>
-        <Stack direction="row" spacing={1} sx={{ alignItems: 'flex-start' }}>
-          <Avatar
-            alt="PlayerHead"
-            src={`https://mc-heads.net/avatar/${entry.authorName}`}
-            sx={{
-              width: 36,
-              height: 36,
-              mt: 0.5,
-              flexShrink: 0,
-              border: isAdmin ? 2 : 0,
-              borderColor: 'success.main',
-            }}
-          />
-          <Stack sx={{ flex: 1 }} spacing={0.25}>
-            <Stack direction="row" spacing={1} sx={{ alignItems: 'center' }}>
-              <Typography
-                variant="subtitle2"
-                component="span"
-                sx={{ fontWeight: 600 }}
-              >
-                {entry.authorName}
-              </Typography>
-              {isAdmin && (
-                <Chip
-                  avatar={
-                    <Avatar
-                      src="/server-icon.png"
-                      sx={{ width: 18, height: 18 }}
-                    />
-                  }
-                  label={capabilities.adminLabel}
-                  color="success"
-                  size="small"
-                  sx={{ height: 20 }}
-                />
-              )}
-              {showDeleteTrigger && (
-                <IconButton
-                  size="small"
-                  color="error"
-                  aria-label="delete"
-                  sx={{ p: 0.5, ml: 'auto' }}
-                  onClick={handleDelete}
-                >
-                  <DeleteIcon fontSize="small" />
-                </IconButton>
-              )}
-            </Stack>
-            <Typography
-              variant="caption"
-              component="span"
-              color="text.secondary"
-            >
-              {formatString(entry.timestamp)}
-            </Typography>
-          </Stack>
-          {showMenuTrigger && (
-            <>
-              <IconButton
-                color="primary"
-                aria-label="その他の操作"
-                onClick={(event: MouseEvent<HTMLElement>) =>
-                  setAnchorEl(event.currentTarget)
-                }
-              >
-                <MoreVert />
-              </IconButton>
-              <Menu
-                anchorEl={anchorEl}
-                open={anchorEl !== undefined}
-                onClose={() => setAnchorEl(undefined)}
-              >
-                {entry.canEdit && (
-                  <MenuItem
-                    onClick={() => {
-                      setIsEditing(true);
-                      setAnchorEl(undefined);
-                    }}
-                  >
-                    編集
-                  </MenuItem>
-                )}
-                {entry.canDelete && (
-                  <MenuItem onClick={handleDelete}>削除</MenuItem>
-                )}
-              </Menu>
-            </>
-          )}
-        </Stack>
+        <ConversationEntryHeader
+          entry={entry}
+          capabilities={capabilities}
+          anchorEl={anchorEl}
+          onOpenMenu={handleOpenMenu}
+          onCloseMenu={handleCloseMenu}
+          onStartEditing={handleStartEditing}
+          onDelete={handleDelete}
+        />
 
         <Box sx={{ pl: '44px', mt: 0.5 }}>
           {isEditing ? (
-            <TextField
-              defaultValue={entry.body}
-              helperText="編集を確定するには Enter キー、キャンセルするには Esc キーを入力してください。"
-              multiline
-              fullWidth
-              onChange={(event) => setDraftBody(event.target.value)}
-              onKeyDown={async (event) => {
-                if (
-                  event.key === 'Enter' &&
-                  !event.shiftKey &&
-                  !event.nativeEvent.isComposing
-                ) {
-                  event.preventDefault();
-                  await handleUpdate();
-                } else if (event.key === 'Escape') {
-                  setDraftBody(entry.body);
-                  setIsEditing(false);
-                }
-              }}
+            <ConversationEntryEditor
+              value={draftBody}
+              onChange={setDraftBody}
+              onSubmit={handleUpdate}
+              onCancel={handleCancelEditing}
             />
           ) : (
-            <Paper
-              variant={entry.renderMode === 'plain' ? 'outlined' : undefined}
-              sx={(theme) => ({
-                p: entry.renderMode === 'plain' ? 1.5 : 0,
-                backgroundColor:
-                  entry.renderMode === 'plain'
-                    ? isAdmin
-                      ? alpha(theme.palette.success.main, 0.08)
-                      : theme.palette.grey[50]
-                    : 'transparent',
-                borderRadius: entry.renderMode === 'plain' ? 2 : 0,
-                boxShadow: 'none',
-              })}
-            >
-              {entry.renderMode === 'markdown' ? (
-                <Box sx={{ whiteSpace: 'pre-wrap' }}>
-                  <Markdown remarkPlugins={[remarkGfm]}>{entry.body}</Markdown>
-                </Box>
-              ) : (
-                <Typography
-                  variant="body2"
-                  component="p"
-                  sx={{ whiteSpace: 'pre-wrap', wordBreak: 'break-word' }}
-                >
-                  {entry.body}
-                </Typography>
-              )}
-            </Paper>
+            <ConversationEntryBody entry={entry} />
           )}
         </Box>
       </Box>
